fix(watchdog): Ventil-Schwelle aus dem Melde-Takt ableiten statt fest 24 h
Wie lange darf ein Ventil schweigen, bevor das Schweigen eine Störung ist? Eine
feste Stundenzahl kann das nicht beantworten: Sie bedeutet für ein Ventil im
5-Minuten-Takt etwas völlig anderes als für eines im Stundentakt. Gemessen an den
Produktionsdaten vom 24.08.2026 verstrichen bei 24 h Schwelle 288 ausgebliebene
Meldungen, bevor der Watchdog etwas sagte — beim zweiten Ventil (6-Sekunden-Takt)
wären es rund 14.400 gewesen.

Für Kameras wendet derselbe Watchdog die richtige Regel längst an
(max(3 × sleep_duration, 1 h)); nur bei Ventilen stand ein geräteunabhängiger
Fixwert. Der Takt ist bei Ventilen nicht konfiguriert, aber aus device_status_log
messbar.

- core/watchdog_threshold.valve_timeout_seconds(): die reine Rechnung
  Faktor × Takt, begrenzt auf [Mindestschwelle, Obergrenze]
- database.get_device_report_interval_seconds(): misst den Takt als Median der
  letzten Abstände. Median statt Mittelwert, weil eine einzelne Lücke —
  typischerweise ein früherer Ausfall — den Mittelwert nach oben zöge und die
  Schwelle genau dann anhöbe, wenn sie greifen soll
- ohne Messdaten bleibt es beim Fixwert: kein Blindflug nach unten
- WATCHDOG_VALVE_TIMEOUT_HOURS bleibt als Obergrenze und gewinnt bei
  widersprüchlicher Konfiguration gegen die Mindestschwelle

Wirkung am realen Ausfall (Rechts Nebelregen, Takt 299 s):
  Schwelle 24.00 h -> 1.00 h
  Alarm bisher 23.08. 14:25 (nach 24.9 h), jetzt 22.08. 14:30 (nach 1.0 h)

Der im Ticket ebenfalls vorgeschlagene zweite Punkt — der Tagesbericht solle das
Alter von last_update zusätzlich selbst bewerten — wurde bewusst nicht umgesetzt.
Er war eine Folge der zu groben Schwelle, nicht ein eigener Mangel; eine zweite
Altersbewertung wäre eine zweite Wahrheit neben dem Flag, und genau diese
Doppelung war die Ursache von telegram_GartenBot-b7y. Begründung im Ticket.

telegram_GartenBot-8zj

+++ src/daemon/core/watchdog_events.py
@@ -2,7 +2,7 @@ from .event_bus import Event
 
 
 class InactivityAlertTriggered(Event):
-    def __init__(self, device_name: str, valve_id: int, hours_silent: float, timeout_hours: int):
+    def __init__(self, device_name: str, valve_id: int, hours_silent: float, timeout_hours: float):
         self.device_name = device_name
         self.valve_id = valve_id
         self.hours_silent = hours_silent
